+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Common & useful classes, resources, extensions. Based on Symfony framework.
 1. Resources > translations > add new
 2. Resources > translations > group by destination in each translation domain
 3. Tests > verify default configuration parameters > add "defaults" environment
+4. Service > FormService > service who serves forms. Configuration > form.novalidate parameter > information if HTML5 
+inline validation is disabled.
 
 # 0.1.4